+++ packages/backend/src/server/api/endpoints/users/notes.ts
@@ -100,7 +100,7 @@ export default class extends Endpoint<typeof meta, typeof paramDef> { // eslint-
 				}
 			}
 
-			if (!this.serverSettings.enableFanoutTimeline) {
+			if (!this.serverSettings.enableFanoutTimeline || !this.serverSettings.fanoutTimelineActive) {
 				const timeline = await this.getFromDb({
 					untilId,
 					sinceId,