+++ packages/misskey-js/src/autogen/types.ts
@@ -9516,6 +9516,7 @@ export interface operations {
                         manifestJsonOverride: string;
                         policies: Record<string, never>;
                         enableFanoutTimeline: boolean;
+                        fanoutTimelineActive: boolean;
                         enableFanoutTimelineDbFallback: boolean;
                         perLocalUserUserTimelineCacheMax: number;
                         perRemoteUserUserTimelineCacheMax: number;